move some common classes to core module - update using of this classes - remove unused class - update using library in application classes
