+++ app/src/main/java/com/github/persiancal/android/MyApplication.kt
@@ -2,8 +2,8 @@ package com.github.persiancal.android
 
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
+import com.github.persiancal.core.CalendarType
 import com.github.persiancal.sdklocal.LocalCalendarEvents
-import com.github.persiancal.sdkremote.CalendarType
 import com.github.persiancal.sdkremote.RemoteCalendarEvents
 
 class MyApplication : Application() {
@@ -17,9 +17,9 @@ class MyApplication : Application() {
             .init(this)
 
         LocalCalendarEvents
-            .addCalendar(com.github.persiancal.sdklocal.CalendarType.JALALI)
-            .addCalendar(com.github.persiancal.sdklocal.CalendarType.HIJRI)
-            .addCalendar(com.github.persiancal.sdklocal.CalendarType.GREGORIAN)
+            .addCalendar(CalendarType.JALALI)
+            .addCalendar(CalendarType.HIJRI)
+            .addCalendar(CalendarType.GREGORIAN)
             .init(this)
     }
 }
