+++ core/src/main/java/com/github/persiancal/core/CalendarType.kt
@@ -1,4 +1,4 @@
-package com.github.persiancal.sdkremote
+package com.github.persiancal.core
 
 enum class CalendarType {
     JALALI,
