--- sdk-local/src/main/java/com/github/persiancal/sdklocal/CalendarType.kt
@@ -1,7 +1,0 @@
-package com.github.persiancal.sdklocal
-
-enum class CalendarType {
-    JALALI,
-    HIJRI,
-    GREGORIAN
-}
